research/Python/requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1623867229539/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1614360979875/work
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
blinker==1.4
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
chardet @ file:///tmp/build/80754af9/chardet_1605303192865/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
coverage @ file:///tmp/build/80754af9/coverage_1614614861224/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1626256956684/work
Django==3.2.6
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
fonttools==4.25.0
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1614884176703/work
h5py @ file:///tmp/build/80754af9/h5py_1608649014695/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877307939/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
Markdown @ file:///tmp/build/80754af9/markdown_1614363852612/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1628003481308/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict @ file:///tmp/build/80754af9/multidict_1607367781142/work
munkres==1.1.4
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264704411/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
opencv-python==4.5.3.56
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pandas==1.3.1
Pillow @ file:///tmp/build/80754af9/pillow_1625655408865/work
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT @ file:///tmp/build/80754af9/pyjwt_1619682484438/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370406642/work
scipy @ file:///tmp/build/80754af9/scipy_1618855955483/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.1
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614360980945/work/tensorflow_pkg/tensorflow-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorflow-estimator_1622026529081/work/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1628668762525/work
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wrapt @ file:///tmp/build/80754af9/wrapt_1607574498026/work
yarl @ file:///tmp/build/80754af9/yarl_1606939947528/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
